fix findbugs errors
